requirements.txt
fastapi==0.110.0
uvicorn[standard]==0.29.0
pydantic==1.10.13
python-multipart==0.0.9
celery==5.3.4
redis==5.0.1
requests==2.31.0
pytest==7.4.4
httpx==0.25.2
spleeter==2.4.0
yt-dlp==2023.12.30
ffmpeg-python==0.2.0
jinja2==3.1.3
python-dotenv==1.0.1
